+++ app/src/main/java/main/tls_maps/placeholder/NotesContent.java
@@ -11,7 +11,7 @@ import java.util.Map;
  * <p>
  * TODO: Replace all uses of this class before publishing your app.
  */
-public class PlaceholderContent {
+public class NotesContent {
 
     /**
      * An array of sample (placeholder) items.
