+++ app/src/main/java/main/tls_maps/ui/notes/Items.java
@@ -13,7 +13,7 @@ import android.view.View;
 import android.view.ViewGroup;
 
 import main.tls_maps.R;
-import main.tls_maps.placeholder.PlaceholderContent;
+import main.tls_maps.placeholder.NotesContent;
 
 /**
  * A fragment representing a list of Items.
@@ -65,7 +65,7 @@ public class Items extends Fragment {
             } else {
                 recyclerView.setLayoutManager(new GridLayoutManager(context, mColumnCount));
             }
-            recyclerView.setAdapter(new MyItemRecyclerViewAdapter(PlaceholderContent.ITEMS));
+            recyclerView.setAdapter(new MyItemRecyclerViewAdapter(NotesContent.ITEMS));
         }
         return view;
     }
